Clarify docker command for macOS
I had issues when running the command on my m2 macbook air. [It turns
out][1] that apparently docker desktop on macOS doesn't really support
the host option.

Additionally, it looks like I had to explicitly pass the port mapping
options for the ports to be actually mapped.

There might be a better and more efficient solution, but this is what I
came up with and it finally worked.

[1]:https://github.com/docker/for-mac/issues/1031