feat(xgroup): support MKSTREAM for xgroup create
Redis supports a MKSTREAM option to dynamically create
a stream for "xgroup create" command if the stream for
the given key doesn't exist. Dragonfly, though instructs
users to use MKSTREAM if a key doesn't exist, doesn't
have the option.

Support MKSTREAM option for "xgroup create" command to
dynamically create a stream for given key, if not already
created.

Signed-off-by: Abhradeep Chakraborty <chakrabortyabhradeep79@gmail.com>