feat: add ScoreMap that maps sds strings to doubles
This map should serve as a replacement for dict that currently is used internally by the SortedMap.
Specifically, it maps sds strings to scores to provide a reverse mapping.
It should be more efficient in memory and CPU - requires a single allocation to store both the
key and the value, in contrast to dict that needs two allocations.

In addition, it also reduces the object size needed for the zset score set:
zsl stores both the sds pointer and its double, i.e. requires 16 bytes not including the metadata.
By passing to bptree_set the internal ScoreMap objects that are allocated by ScoreMap::AddOrUpdate
we can just store 8-byte pointers, not including the bptree metadata.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>