fix: possible segfault bug
A user reported crash in Connection::GetClientInfo function.
When connection just start running, its phase_ variable was not
initialized, so a possible "CLIENT LIST" command from another connection
could trigger invalid data access when printing phase_.
To be on the safe side I also verify that other pointers in the function
are set (could also be that we have a similar issue during connection shutdown).

Signed-off-by: Roman Gershman <roman@dragonflydb.io>