bug(server): zadd wrong insert when non unique members
the bug: when calling zadd with non uniq members the server would result
in updating the zset with the higher score instead of the last score
given
f.e
zadd myzset 2 a 1 a
will result in member a score 2 while the expected 1
the fix: if the members in zadd command are not unique than we do not
sort the members. As a result in this case when we dont sort we will not have perfomace optimization on listpack.

Signed-off-by: adi_holden <adi@dragonflydb.io>