bug(server): global command stalls on server load with pipeline mode
the bug: global command is not able to schedule into txq when stress
testing the server using pipeline mode. Only after the stress test finish the global
trasaction gets scheduled into the txq. The reason for this is when we
start a global transaction we set the shard lock and all the
transactions start to enter the txq. They compete with the global tx on
the order they are inserted into the queue to preserve transaction
atomicity. Because the global tx needs to be inserted to all shard queues
its chance to schedule with order with all the other trasactions is low.
the solution: lock the global transaction inside the schedual in shard,
locking closer to schedualing decreases the number of trasactions in the queue
and the compatition on oredering correctly has higher chance now.

Signed-off-by: adi_holden <adi@dragonflydb.io>