chore: flip default sorted set implementation to bptree
Also, bring back the default max listpack entries count for zset to 128.
The reason for this - I've added some optimizations that improved listpack
performance and also because I would like to write an article about it
and I need to compare Dragonfly to Redis7 that has this setting set to 128
by default.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>