feat(makefile): Build with full debug symbols and strip in `make package`
Before this PR, using `make` for building releases used `-g1 -gz` to
provide some level of debug symbols (mostly function names in stack
traces).

We want to be able to have binaries with full debug symbols (for
coredumps etc), but not in the public releases.

This PR builds with full debug symbols, but changes `make package` to
strip most debug symbols, while keeping `.debug_line` (and keeping them
compressed as if with `-gz`). This will allow a single build command for
public releases as well as releases with full symbols.