chore: allow aggregate functions for jsonpath (#2611)
Specifically, allows expressions in the form of
`max($.objs[*].score`, i.e. where a function aggregates 0 or more values
produced by the path expression. The final value is passed to a PathCallback.
In case no matches are found, json null value is passed. In any case, the callback
is called exactly once for such expressions.

In addition, fixed a parser bug that does not allow fields after the bracketing expression.

Finally, introduced an utility function for easier parsing of json::Path.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>