chore: add SBF data structure
Based on https://gsd.di.uminho.pt/members/cbm/ps/dbloom.pdf

The data-structure itself is a growing list of bloom filters,
where the next filter has exponentially larger capacity with exponentially tighter error bound.

The Exist() goes over all the filters and it's enough that at least one of them returns a positive result.
For Add(), we make ensure that all the existing filters do not have the element, as well as making sure that the last
filter that is being filled does not cross its maximum designated capacity.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>