feat: extended bracket index in jsonpath
Before this change, we only allowed a single non-negative index or a wildcard inside brackets [].
For example, "$.foo[1]".
This PR extends the indexing expression to negative indices. It also builds up the foundation
to support [begin:end] slices, though this has not been implemented in this PR.

Finally, it adds more coverage to tests.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>