chore: allow calling Context::ReportError from I/O dispatch fiber
Fixes the issue when we are trying to Report the error from JournalStreamer::OnCompletion
that is called from a proactor callback directly from the I/O fiber.

We can not use fiber mutexes in the I/O fiber so we switch to regular mutex and reduce its critical sections
as much as possible.

Should fix the following failure:
https://github.com/dragonflydb/dragonfly/actions/runs/9536876957/job/26284317328#step:6:622

Signed-off-by: Roman Gershman <roman@dragonflydb.io>