chore: Separate tiered serialization format from object values
Before that OpManager::Stash, DiskStorage::Stash and SmallBins::Stash received a
single value argument as a serialized blob of an object value.

Now they accept an additional footer argument that allows appending arbitrary byte sequence when
preparing a disk page. We could, of course serialize, into value before calling these interface but then
we would need another copy.
With two arguments we allow an optional meta-format that has a footer after the value.

This PR does not use this feature yet, but we will need it in the next PR in order to store the encoding mask of the string object.
In the future we will need storing other meta information in case we decide supporting multiple DF types.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>