fix: Do not publish to connections without context
This is a rare case where a closed connection is kept alive while the
handling fiber yields, therefore leaving `cc_` (the connection context)
pointing to null for other fibers to see.

As far as I can see, this can only happen during server shutdown, but
there could be other cases that I have missed.

The test on its own does _not_ reproduce the crash, however with added
`ThisFiber::SleepFor()`s I could reproduce the crash:

* Right before `DispatchBrief()`
  [here](https://github.com/dragonflydb/dragonfly/blob/e3214cb603fce42c8bb52a2c220c3115b7f8cecf/src/server/channel_store.cc#L154)
* Right after connection context `reset()`
  [here](https://github.com/dragonflydb/dragonfly/blob/2ab480e160dece516b908584f1933b25b760abfb/src/facade/dragonfly_connection.cc#L750)

In any case, calling `SendPubMessageAsync()` to a connection where `cc_`
is null is a bug, and we fix that here.